caam: Re-apply HW Scatter/Gather Assigment Fix for ARM Platforms
Since the CAAM driver was originally written for a BE (Power) platform,
the original definition of the hardware scatter/gather list was
incompatible with LE (ARM) platforms, and therefore was redefined for
both platforms in the i.MX branch during the original 2012 port for
i.MX6. Some of these changes were introduced in a patch entitled
"Rework scatterlist handling for bi-endian platforms" and committed
on 05-July-2012.

The very last change in this patch amended an inline population function
such that the length assignment would OR-preserve the co-resident E
(extended) or F (final entry) bits if used. In the 2 years hence, the
single change had been lost. Thererfore, this patch reintroduces the
necessary change.

Signed-off-by: Steve Cornelius <steve.cornelius@freescale.com>